refactor(document,ui): Move TreeDiagram from to ui/ package
