+++ src/main/java/net/boyechko/pdf/autoa11y/ui/cli/PdfAutoA11yCLI.java
@@ -39,12 +39,12 @@ import net.boyechko.pdf.autoa11y.core.ProcessingService;
 import net.boyechko.pdf.autoa11y.document.Content;
 import net.boyechko.pdf.autoa11y.document.PdfCustodian;
 import net.boyechko.pdf.autoa11y.document.StructTree;
-import net.boyechko.pdf.autoa11y.document.TreeDiagram;
 import net.boyechko.pdf.autoa11y.ui.AccessibilityReport;
 import net.boyechko.pdf.autoa11y.ui.FormattedListener;
 import net.boyechko.pdf.autoa11y.ui.LoggingListener;
 import net.boyechko.pdf.autoa11y.ui.SidecarConfig;
 import net.boyechko.pdf.autoa11y.ui.StructTreeTablePrinter;
+import net.boyechko.pdf.autoa11y.ui.TreeDiagram;
 import net.boyechko.pdf.autoa11y.ui.VerbosityLevel;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
